requirements.txt
azure-functions
requests==2.23.0